Big changes (Ship class, firing, hit/miss markers)
Features:
Added class system for ships so theu can be more easily target and more easily see each ships information.

Added ship hit detection (only detects enemy side only atm)
If it's a hit, it add's a hitmarker on the spot.
If it's a miss, it add's a missmarker on the spot.

Bugs:
When generating enemy ships sometimes it crashes. Maybe add sleep time so it's slower?